Add lib/types.py
