+++ lib/__init__.py
@@ -3,11 +3,12 @@ from __future__ import annotations
 import logging
 
 from .io import load_json, save_json
-from .item import Item, item_dict
+from .item import item_dict
 from .merge import merge
-from .servant import Servant, servant_dict, servant_list, skill_dict
+from .servant import servant_dict, servant_list, skill_dict
 from .sound import Sound, sound_list
 from .text import Dictionary, Text
+from .types import Item, Servant
 from .validate import validate_servant, validate_servants
 
 logging.getLogger(__package__).addHandler(logging.NullHandler())
